+++ esp_data/dataset/webds.py
@@ -6,10 +6,10 @@ from typing import Any, Callable, Generator, Literal, Union
 import pandas as pd
 import webdataset as wds
 
-import esp_data.file_io.functional as F
+import esp_data.io.functional as F
 from esp_data.config import DatasetConfig
 from esp_data.config.project_config import default_webds_loader_cfg
-from esp_data.paths import AnyPath
+from esp_data.io import AnyPath
 
 from .base import BaseIterableDataset, BaseMapDataset
 from .shard_creator import write_webdataset_shard
