+++ esp_data/io/__init__.py
@@ -1,0 +1,3 @@
+from esp_data.io.paths import AnyPath, GSPath, R2Path
+
+__all__ = ["AnyPath", "GSPath", "R2Path"]
